requirements.txt
# Core dependencies
beautifulsoup4>=4.11.1
soupsieve>=2.0
urllib3[socks]>=1.26.11
pynput==1.7.7
lxml
Xlib
pywin32>=303; sys_platform == "win32"
distro>=1.8.0
semver>=3.0.2
zstandard>=0.23.0
send2trash>=1.8.3
webdav4==0.10.0
dateparser==1.2.1
pyside6-essentials==6.10.1
markdown>=3.7

# Transitive dependencies
regex
tzlocal
pytz
python-dateutil
